+++ lib/fork.js
@@ -30,6 +30,10 @@ if (env.NODE_PATH) {
 		.join(path.delimiter);
 }
 
+// In case the test file imports a different AVA install, the presence of this variable allows it to require this one
+// instead.
+env.AVA_PATH = path.resolve(__dirname, '..');
+
 module.exports = function (file, opts, execArgv) {
 	opts = objectAssign({
 		file: file,
